% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freefall.R
\name{freefall}
\alias{freefall}
\title{calculate the distance traveled by a falling object
using the equation of motion h = 1/2*gt^2, return the time
it takes for the free fall.}
\usage{
freefall(height, g = 9.8)
}
\arguments{
\item{height}{A numeric value that is the distance of the free fall in meters.}

\item{g}{The gravity of the free fall.}
}
\value{
time A number that is the time it takes for the free fall.
}
\description{
calculate the distance traveled by a falling object
using the equation of motion h = 1/2*gt^2, return the time
it takes for the free fall.
}
\examples{
freefall(10, g = 9.8)
1.4286
freefall(10, g = 274)
0.2702

}
