% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetic_energy.R
\name{kinetic_energy}
\alias{kinetic_energy}
\title{Kinetic Energy}
\usage{
kinetic_energy(m, v)
}
\arguments{
\item{m}{mass of the object (kg)}

\item{v}{velocity of the object (m/s)}
}
\value{
KE kinetic energy of the object
}
\description{
Calculates the Kinetic Energy of an object. When work is done on an object, energy is transferred, and the object moves with a new constant speed. We call the energy that is transferred kinetic energy, and it depends on the mass and speed achieved. The kinetic energy equation is given as: KE = 1/2\emph{m}v^2, where KE is the kinetic energy, m is the body’s mass, and v is the body’s velocity.
}
\examples{
m <- 0.6
v <- 3
kinetic_energy(m, v)
2.7
}
