% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_friction_ground.R
\name{static_friction_ground}
\alias{static_friction_ground}
\title{calculating the friction force for static object. The formula is fr = mu * N, where
the mu is the coefficient of friction which incorporating the characteristics of the surface.
For example, if you want to calculate the friction when you put your phone on a table, then the coefficient of friction is
depending on the matirial of the table. N is the normal force which means the perpendicular force of the surface which object
touching on. In our case, it would be m*g since it is on the ground. fr = mu * m * g.}
\usage{
static_friction_ground(mu, m, g = 9.8)
}
\arguments{
\item{mu}{A numeric value that is the coefficient of friction}

\item{m}{A numeric value that is the mass of the object}

\item{g}{The gravity of the free fall.}
}
\value{
the static friction force
}
\description{
calculating the friction force for static object. The formula is fr = mu * N, where
the mu is the coefficient of friction which incorporating the characteristics of the surface.
For example, if you want to calculate the friction when you put your phone on a table, then the coefficient of friction is
depending on the matirial of the table. N is the normal force which means the perpendicular force of the surface which object
touching on. In our case, it would be m*g since it is on the ground. fr = mu * m * g.
}
\examples{
static_friction_ground(mu=0.2, g = 9.8, m=10)
19.6
freefall(g=10, mu=0.2, m=2)
4

}
