% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freefall.R, R/freefall_plot.R
\name{freefall}
\alias{freefall}
\title{calculate the time it takes for a free fall based on
distance traveled by a falling object and the gravity of the free fall
using the equation of motion h = 1/2*gt^2. It also returns
a plot compares the time it takes for the free fall on different planets.}
\usage{
freefall(height, g = 9.8)

freefall(height, g = 9.8)
}
\arguments{
\item{height}{A numeric value that is the distance of the free fall in meters.}

\item{g}{The gravity of the free fall.}
}
\value{
time A number that is the time it takes for the free fall.
plot A plot compares the time of free fall on different planets.

time A number that is the time it takes for the free fall.
plot A plot compares the time of free fall on different planets.
}
\description{
calculate the time it takes for a free fall based on
distance traveled by a falling object and the gravity of the free fall
using the equation of motion h = 1/2*gt^2. It also returns
a plot compares the time it takes for the free fall on different planets.

calculate the time it takes for a free fall based on
distance traveled by a falling object and the gravity of the free fall
using the equation of motion h = 1/2*gt^2. It also returns
a plot compares the time it takes for the free fall on different planets.
}
\examples{
freefall(10, g = 9.8)
1.4286
freefall(10, g = 274)
0.2702

freefall(10, g = 9.8)
1.4286
freefall(10, g = 274)
0.2702

}
